refactor(TabBase): Make reusable
Refactored reused partials
