+++ packages/react-component-library/src/components/TabBase/partials/StyledTabs.tsx
