+++ packages/react-component-library/src/components/TabNav/TabNav.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { Nav, NavItem } from '../../common/Nav'
 import { StyledTabNav } from './partials/StyledTabNav'
-import { StyledTabs } from '../TabSet/partials/StyledTabs'
+import { StyledTabs } from '../TabBase/partials/StyledTabs'
 
 export const TabNav: React.FC<Nav<NavItem>> = ({ children, className }) => (
   <StyledTabNav className={className}>
