+++ packages/react-component-library/src/components/TabNav/TabNavItem.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
 import { NavItem } from '../../common/Nav'
-import { StyledTabItem } from '../TabSet/partials/StyledTabItem'
-import { StyledTab } from '../TabSet/partials/StyledTab'
+import { StyledTabItem } from '../TabBase/partials/StyledTabItem'
+import { StyledTab } from '../TabBase/partials/StyledTab'
 
 export const TabNavItem: React.FC<NavItem> = ({ isActive, link }) => (
   <StyledTabItem data-testid="tab-nav-tab">
