+++ packages/react-component-library/src/components/TabSet/TabSet.tsx
@@ -10,7 +10,7 @@ import { StyledHeader } from './partials/StyledHeader'
 import { StyledNavigation } from './partials/StyledNavigation'
 import { StyledScrollLeft } from './partials/StyledScrollLeft'
 import { StyledScrollRight } from './partials/StyledScrollRight'
-import { StyledTabs } from './partials/StyledTabs'
+import { StyledTabs } from '../TabBase/partials/StyledTabs'
 import { StyledTabSet } from './partials/StyledTabSet'
 import { TabSetItem, TabContent, TabItem, TabSetItemProps } from '.'
 import { useScrollableTabSet } from './useScrollableTabSet'
